requirements.txt
fastapi
uvicorn
sounddevice
numpy
whisper
torch
opencc-python-reimplemented
python-multipart
aiofiles